+++ collector-bitmovin-player/src/androidTest/java/com/bitmovin/analytics/bitmovin/player/ErrorScenariosTest.kt
@@ -21,8 +21,6 @@ import com.bitmovin.player.api.Player
 import com.bitmovin.player.api.PlayerConfig
 import com.bitmovin.player.api.TweaksConfig
 import com.bitmovin.player.api.analytics.create
-import com.bitmovin.player.api.network.HttpRequestType
-import com.bitmovin.player.api.network.NetworkConfig
 import com.bitmovin.player.api.source.Source
 import com.bitmovin.player.api.source.SourceConfig
 import kotlinx.coroutines.MainScope
@@ -33,8 +31,6 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import java.util.concurrent.CompletableFuture
-import java.util.concurrent.TimeUnit
 
 @RunWith(AndroidJUnit4::class)
 class ErrorScenariosTest {
@@ -164,7 +160,7 @@ class ErrorScenariosTest {
             assertThat(errorDetail.data.exceptionMessage).isNullOrEmpty()
             assertThat(errorDetail.data.additionalData).isNotNull()
             // we verify that additional data is not truncated too early
-            assertThat(errorDetail.data.additionalData).hasSizeGreaterThan(2000)
+            assertThat(errorDetail.data.additionalData).hasSizeGreaterThan(1500)
             // sanity check that additional data is not cutoff in the middle of a json object
             // This is rudimentary, but should catch most cases
             assertThat(errorDetail.data.additionalData).endsWith("}")
@@ -212,75 +208,5 @@ class ErrorScenariosTest {
             DataVerifier.verifyStaticErrorDetails(errorDetail, impressionId, defaultAnalyticsConfig.licenseKey)
             assertThat(errorDetail.data.exceptionStacktrace).isNotEmpty()
             assertThat(errorDetail.data.exceptionMessage).isNotEmpty()
-        }
-
-    @Test
-    fun test_liveStreamGettingBehindLiveWindow_Should_sendErrorSample() =
-        runBlockingTest {
-            val stream = Samples.DASH_LIVE
-            val sourceMetadata =
-                SourceMetadata(
-                    title = metadataGenerator.getTestTitle(),
-                    customData = CustomData(customData1 = "nonExistingStream"),
-                )
-
-            val source = Source.create(SourceConfig.fromUrl(stream.uri.toString()), sourceMetadata)
-            // act
-            var index = 0
-
-            withContext(mainScope.coroutineContext) {
-                val playerConfig =
-                    PlayerConfig(
-                        key = "a6e31908-550a-4f75-b4bc-a9d89880a733",
-                        playbackConfig = PlaybackConfig(),
-                        tweaksConfig = TweaksConfig(enableMainContentVideoCodecInitializationFallback = false),
-                    )
-                val player = Player.create(appContext, playerConfig, defaultAnalyticsConfig)
-
-                // network config to simulate a slow network
-                // which should cause a BehindLiveWindowException
-                player.config.networkConfig =
-                    NetworkConfig(
-                        preprocessHttpResponseCallback = { type, response ->
-                            if (type == HttpRequestType.MediaVideo) {
-                                // allow the first requests and slow down the others
-                                // which causes a BehindLiveWindowException
-                                index++
-                                if (index > 4) {
-                                    Thread.sleep(index * 3000L)
-                                }
-                                CompletableFuture.completedFuture(response)
-                            } else {
-                                CompletableFuture.completedFuture(response)
-                            }
-                        },
-                    )
-
-                player.load(source)
-                player.play()
-            }
-
-            waitForErrorDetailSample(100, TimeUnit.SECONDS)
-
-            // assert
-            val impressionList = MockedIngress.waitForRequestsAndExtractImpressions()
-            assertThat(impressionList.size).isEqualTo(1)
-
-            val impression = impressionList.first()
-            assertThat(impression.errorDetailList.size).isEqualTo(1)
-
-            val eventDataWithError = impression.eventDataList.find { e -> e.errorCode != null }
-            assertThat(eventDataWithError).isNotNull
-            eventDataWithError!!
-            val errorDetail = impression.errorDetailList.first()
-
-            assertThat(
-                eventDataWithError.errorMessage,
-            ).startsWith("A general error occurred:")
-            assertThat(eventDataWithError.errorCode).isEqualTo(2001)
-
-            val impressionId = eventDataWithError.impressionId
-            DataVerifier.verifyStaticErrorDetails(errorDetail, impressionId, defaultAnalyticsConfig.licenseKey)
-            assertThat(errorDetail.data.exceptionStacktrace.toString()).contains("BehindLiveWindowException")
         }
 }